Refactor: move utils models to data.models
