+++ app/src/main/java/fr/free/nrw/commons/data/models/utils/utils/ConnectionType.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.utils.model
+package fr.free.nrw.commons.data.models.utils.utils
 
 enum class ConnectionType(private val text: String) {
     WIFI_NETWORK("wifi"), CELLULAR_4G("cellular-4g"), CELLULAR_3G("cellular-3g"), CELLULAR("cellular"), NO_INTERNET("no-internet");
