+++ app/src/main/java/fr/free/nrw/commons/data/models/utils/utils/NetworkConnectionType.kt
@@ -1,4 +1,4 @@
-package fr.free.nrw.commons.utils.model
+package fr.free.nrw.commons.data.models.utils.utils
 
 enum class NetworkConnectionType {
     WIFI, TWO_G, THREE_G, FOUR_G, UNKNOWN
